backend/time-series-forecasting/requirements.txt
fastapi
humanize
uvicorn[standard]
pandas
prophet
google-cloud-bigquery[pandas]
google-cloud-aiplatform
pyarrow
numpy

